refactor: 로그 탭에 이전 메인 피드 모달 UI 적용
